HV-1487 Avoid some code duplication
