+++ engine/src/main/java/org/hibernate/validator/internal/engine/resolver/CachingJPATraversableResolverForSingleValidation.java
@@ -56,50 +56,10 @@ class CachingJPATraversableResolverForSingleValidation implements TraversableRes
 		return true;
 	}
 
-	private static final class TraversableHolder {
-
-		private final Object traversableObject;
-		private final Path.Node traversableProperty;
-		private final int hashCode;
+	private static class TraversableHolder extends AbstractTraversableHolder {
 
 		private TraversableHolder(Object traversableObject, Path.Node traversableProperty) {
-			this.traversableObject = traversableObject;
-			this.traversableProperty = traversableProperty;
-			this.hashCode = buildHashCode();
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if ( this == o ) {
-				return true;
-			}
-			if ( o == null || TraversableHolder.class != o.getClass() ) {
-				return false;
-			}
-
-			TraversableHolder that = (TraversableHolder) o;
-
-			if ( traversableObject != null ? !traversableObject.equals( that.traversableObject ) : that.traversableObject != null ) {
-				return false;
-			}
-			if ( !traversableProperty.equals( that.traversableProperty ) ) {
-				return false;
-			}
-
-			return true;
-		}
-
-		@Override
-		public int hashCode() {
-			return hashCode;
-		}
-
-		public int buildHashCode() {
-			// HV-1013 Using identity hash code in order to avoid calling hashCode() of objects which may
-			// be handling null properties not correctly
-			int result = traversableObject != null ? System.identityHashCode( traversableObject ) : 0;
-			result = 31 * result + traversableProperty.hashCode();
-			return result;
+			super( traversableObject, traversableProperty );
 		}
 	}
 }
